Fix bug where user needs to login twice
Because MoodsProvider wrapped all routes, it was loaded before the TokenHandler
could set the token in local storage and would redirect the user back to the login
component as an unauthed user. Added a second Switch component for journal routes
and only wrap those in MoodsProvider so that there is no issue with API calls
and when the JWT is set in localStorage.